feat(snap): full-precision snap→commit path (REQ-101 ±0.002 ft, ADR-054 Phase C, #442)
TASK-228 Phase C. The object-snap → pick-commit chain is now `double` end to end,
so REQ-101's "a snapped pick is bit-identical to the vertex it snapped to"
property holds at any drawing magnitude — a `float` copy on that path broke it
above ~10,000 ft local.

- CadSnap::Hit::x/y/z → double.
- AppCommandState::viewportSnapPickLocalX/Y/Z → double.
- SubmitViewportPick / SubmitViewportPickImpl / UiSubmitViewportPick entry
  coordinates → double; ApplySegmentAnglePickToViewportPick → double&.
- CadUi commitX/commitY and the snapped curMX/curMY submit calls keep double
  (dropped their now-pointless static_cast<float>).

The ORTHO / polar / angle-lock constraint helpers keep working in `float`
(narrowed across the call and back): an axis/ray-locked pick is a computed,
pixel-bounded point that REQ-101 explicitly scopes out of ±0.002 ft, and it is
mutually exclusive with an active object snap. Rubber-band preview buffers stay
`float` — render-only, GPU-bound.

Only 12 boundary sites needed edits; the `float wx/wy` pick handlers compile
unchanged (double→float at their internal comparisons, warning-suppressed).

Build clean; ctest 1359/1359.

+++ src/viewport/CadSnap.hpp
@@ -48,12 +48,12 @@ enum class Kind {
 struct Hit {
   bool valid = false;
   Kind kind = Kind::Endpoint;
-  float x = 0.f;
-  float y = 0.f;
+  double x = 0.0;
+  double y = 0.0;
   /// Elevation of the snapped point (REQ-057/058). Without it the snap glyph is drawn on the
   /// datum while the point it marks sits at its own elevation, so the marker floats away from the
   /// geometry as soon as the view is orbited. Zero for flat drawings, which is every pre-3D one.
-  float z = 0.f;
+  double z = 0.0;
   /// True when this candidate came from the 3D Object Snap system (REQ-325/#395) rather than the 2D
   /// one — used only to color the glyph (3D Object Snap purple #8803fc vs 2D Object Snap green), so
   /// a user can tell at a glance which snap system answered, even for a kind (Endpoint, Midpoint,
